api woriking

+++ src/modules/vehicles/vehicle.service.ts
@@ -1,5 +1,13 @@
 import { pool } from "../../config/db";
 
+interface VehiclePayload {
+  vehicle_name?: string;
+  type?: string;
+  registration_number?: string;
+  daily_rent_price?: number;
+  availability_status?: string;
+}
+
 const createVehicle = async (payload: Record<string, unknown>) => {
   const {
     vehicle_name,
@@ -9,11 +17,13 @@ const createVehicle = async (payload: Record<string, unknown>) => {
     availability_status,
   } = payload;
 
+  const vehicleType = type || "car";
+
   const result = await pool.query(
     `INSERT INTO vehicles(vehicle_name, type, registration_number, daily_rent_price, availability_status) VALUES($1, $2, $3, $4, $5) RETURNING *`,
     [
       vehicle_name,
-      (type as string).toLowerCase(),
+      (vehicleType as string).toLowerCase(),
       registration_number,
       daily_rent_price,
       availability_status,
@@ -34,30 +44,48 @@ const getSingleVehicle = async (id: string) => {
   return result;
 };
 
-const updateVehicle = async (
-  vehicle_name: string,
-  type: string,
-  registration_number: string,
-  daily_rent_price: number,
-  availability_status: string,
-  id: string
-) => {
+const updateVehicle = async (vehicleId: string, payload: VehiclePayload) => {
+  const keys = Object.keys(payload) as (keyof VehiclePayload)[];
+
+  if (keys.length === 0) {
+    throw new Error("No fields provided for update");
+  }
+
+  const setQuery = keys
+    .map((key, index) => `${key} = $${index + 1}`)
+    .join(", ");
+  const values = keys.map((key) => {
+    if (key === "type" && payload[key]) {
+      return payload[key]!.toLowerCase();
+    }
+    return payload[key];
+  });
+
+  values.push(vehicleId);
+
   const result = await pool.query(
-    `UPDATE vehicles SET vehicle_name = $1, type = $2, registration_number = $3, daily_rent_price = $4, availability_status = $5 WHERE id = $6 RETURNING *`,
-    [
-      vehicle_name,
-      type.toLowerCase(),
-      registration_number,
-      daily_rent_price,
-      availability_status,
-      id,
-    ]
+    `UPDATE vehicles SET ${setQuery} WHERE id = $${
+      keys.length + 1
+    } RETURNING *`,
+    values
   );
 
   return result;
 };
 
 const deleteVehicle = async (id: string) => {
+  const activeBookingCheck = await pool.query(
+    `SELECT * FROM bookings 
+     WHERE vehicle_id = $1 AND status = 'active'`,
+    [id]
+  );
+
+  if (activeBookingCheck.rows.length > 0) {
+    throw new Error(
+      "Vehicle cannot be deleted because it has active bookings!"
+    );
+  }
+
   const result = await pool.query(
     `DELETE FROM vehicles WHERE id = $1 RETURNING *`,
     [id]
